clean-up of privateQueues list manipulation functions and descriptor WhichQueue values. Here's a summary:
PostedSpecificRecv[src] - protected by recvLock[src]
PostedWildRecv - protected by list lock
MatchedRecv[src] - protected by list lock
AheadOfSeqRecvFrags[src] - protected by recvLock[src]
OkToMatchRecvFrags[src] - protected by recvLock[src]
OkToMatchSMPFrags[src] - protected by recvLock[src]

+++ src/queue/Pt2PtGroupProgress.cc
@@ -149,7 +149,7 @@ int push_frags_into_network(double timeNow)
                     nAcked = SendDesc->pathInfo.sharedmem.sharedData->NumAcked;
                 }
                 if ( ( nAcked >= SendDesc->numfrags) &&
-                	 (SendDesc->NumSent >= SendDesc->numfrags) ) {
+                	 (SendDesc->NumSent >= (int)SendDesc->numfrags) ) {
                     // message has been acked
                     SendDesc_t *TmpDesc = (SendDesc_t *)
                         IncompletePostedSends.RemoveLinkNoLock(SendDesc);
